requirements.txt
PySide6>=6.6.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
bcrypt>=4.0.0
qrcode>=7.0
reportlab>=4.0.0
pillow>=10.0.0
python-dateutil>=2.8.0
pytest>=7.0.0
matplotlib>=3.7.0
psutil>=5.9.0
PyInstaller>=6.0.0
requests>=2.31.0
